+++ tests/Commands/Server/ToolOperationsTest.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using System.Text.Json;
 using AzureMcp.Commands;
 using AzureMcp.Commands.Server;
 using Microsoft.Extensions.DependencyInjection;
@@ -13,6 +14,18 @@ using Xunit;
 namespace AzureMcp.Tests.Commands.Server;
 public class ToolOperationsTest
 {
+    // https://json-schema.org/understanding-json-schema/reference/type
+    private static readonly HashSet<string> JsonSchemaDataTypes = new()
+    {
+        "string",
+        "integer",
+        "number",
+        "boolean",
+        "array",
+        "null",
+        "object"
+    };
+
     private readonly CommandFactory _commandFactory;
     private readonly IServiceProvider _serviceProvider;
     private readonly ILogger<ToolOperations> _logger;
@@ -48,6 +61,31 @@ public class ToolOperationsTest
             Assert.NotNull(tool);
             Assert.NotNull(tool.Name);
             Assert.NotNull(tool.Description!);
+
+            Assert.Equal(JsonValueKind.Object, tool.InputSchema.ValueKind);
+
+            foreach (var properties in tool.InputSchema.EnumerateObject())
+            {
+                if (properties.NameEquals("type"))
+                {
+                    Assert.Equal("object", properties.Value.GetString());
+                }
+
+                if (!properties.NameEquals("properties"))
+                {
+                    continue;
+                }
+
+                var commandArguments = properties.Value.EnumerateObject().ToArray();
+                foreach (var argument in commandArguments)
+                {
+                    var argumentType = argument.Value.GetProperty("type");
+                    var value = argumentType.GetString();
+
+                    Assert.NotNull(value);
+                    Assert.Contains(value, JsonSchemaDataTypes);
+                }
+            }
         }
     }
 }